Extract cluster edit locic to component
